Rename opsim directory to obstable
