+++ docs/glossary.rst
@@ -19,9 +19,9 @@ Glossary
 
 **Observer Frame**: The reference frame of the observer. Observations in the observer frame account for effects such as redshift.
 
-**ObsTable**: The ObsTable represents the set of data about the individual observations being simulated, including where the telescope is pointing (RA, Dec) and conditions affecting the detector noise.
+**ObsTable**: The ``ObsTable`` represents the set of data about the individual observations being simulated, including where the telescope is pointing (RA, Dec) and conditions affecting the detector noise.
 
-**OpSim**: A specific version of the ObsTable that stores data for the Rubin Observatory's LSST `Operations Simulator <https://www.lsst.org/scientists/simulations/opsim>`_ outputs.
+**OpSim**: A specific version of the ``ObsTable`` that stores data for the Rubin Observatory's LSST `Operations Simulator <https://www.lsst.org/scientists/simulations/opsim>`_ outputs.
 
 **Parameter**: A parameter within the model corresponds to a variable in the mathematical equations for simulating flux. Each run of the simulation effectively samples the parameters in the model and uses them to compute the flux for a given model. Parameters may themselves be the result of computations performed with other parameters. For example we may choose an object's brightness parameter from a Gaussian distribution that is parameterized by mean and standard deviation. Parameters are generated by nodes (``ParameterizedNode`` objects) and stored in the ``GraphState`` objects.
