+++ tests/tdastro/conftest.py
@@ -49,7 +49,7 @@ def test_conditional_flow_filename(test_data_dir):
 @pytest.fixture
 def oversampled_observations(opsim_shorten):
     """Return an OpSim object with 0.05 day cadence spanning year 2027."""
-    from tdastro.opsim.opsim import OpSim, oversample_opsim
+    from tdastro.obstable.opsim import OpSim, oversample_opsim
 
     base_opsim = OpSim.from_db(opsim_shorten)
     return oversample_opsim(
